// packages/shared/src/types/chart-token.ts
export enum ChartTokenId {
  HRV_7DAYS = 'HRV_7DAYS',
  SLEEP_7DAYS = 'SLEEP_7DAYS',
  RESTING_HR_7DAYS = 'RESTING_HR_7DAYS',
  ACTIVITY_7DAYS = 'ACTIVITY_7DAYS',
  SPO2_7DAYS = 'SPO2_7DAYS',
  SLEEP_STAGE_LAST_NIGHT = 'SLEEP_STAGE_LAST_NIGHT',
  STRESS_LOAD_7DAYS = 'STRESS_LOAD_7DAYS',
  HRV_SLEEP_14DAYS_COMPARE = 'HRV_SLEEP_14DAYS_COMPARE',
}